+++ src/main/java/org/discordbots/api/client/DiscordBotListAPI.java
@@ -10,7 +10,14 @@ import java.time.format.DateTimeFormatter;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 import java.util.concurrent.CompletionStage;
-import okhttp3.*;
+import okhttp3.Call;
+import okhttp3.Callback;
+import okhttp3.HttpUrl;
+import okhttp3.MediaType;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.RequestBody;
+import okhttp3.Response;
 import org.discordbots.api.client.entity.PaginatedVotes;
 import org.discordbots.api.client.entity.PartialVote;
 import org.discordbots.api.client.entity.Project;
