+++ src/webhooks/java/org/discordbots/webhooks/IntegrationCreatePayload.java
@@ -1,4 +1,4 @@
-package org.discordbots.api.client.webhooks;
+package org.discordbots.webhooks;
 
 import com.google.gson.annotations.SerializedName;
